@@ -10,6 +10,7 @@
     "build:dist": "shx rm -rf build-dist && tsc -b tsconfig.dist.json && yarn copy:dist",
     "copy:build": "ulx-repo-after-build && ulx-repo-adjust-paths && node alsoCopy.js",
     "copy:dist": "ulx-repo-after-build --dist && shx cp -r \"browser-emulator-builder/data\" \"build-dist/plugins/default-browser-emulator\" && ulx-repo-adjust-paths build-dist",
+    "clean": "tsc -b --clean",
     "da:start": "cd build && yarn workspace @double-agent/collect-controller start",
     "tsc": "tsc -b -i tsconfig.json && yarn copy:build",
     "watch": "yarn tsc && tsc -b -w tsconfig.json",
@@ -19,14 +20,11 @@
     "test:debug:fast": "yarn tsc && cd ./build && cross-env ULX_DATA_DIR=.data-test NODE_ENV=test node --inspect node_modules/.bin/jest --runInBand",
     "lint": "eslint --cache ./",
     "version:check": "ulx-repo-version-check fix",
-    "version:bump": "ulx-repo-version-bump"
+    "version:bump": "ulx-repo-version-bump",
+    "publish:all": "yarn workspaces foreach --topological-dev npm publish --access public"
   },
   "repository": "git@github.com:ulixee/unblocked.git",
   "author": "Ulixee",
-  "contributors": [
-    "Caleb Clark",
-    "Blake Byrnes"
-  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/ulixee/unblocked/issues"
@@ -38,32 +36,19 @@
   "devDependencies": {
     "@commitlint/cli": "^18.6.0",
     "@commitlint/config-conventional": "^18.6.0",
+    "@lerna-lite/version": "^3.9.1",
     "@types/jest": "^29.5.11",
     "@types/node": "^18.19.10",
-    "@typescript-eslint/eslint-plugin": "^6.19.1",
-    "@typescript-eslint/parser": "^6.19.1",
-    "@ulixee/repo-tools": "^1.0.31",
+    "@ulixee/repo-tools": "^1.0.33",
     "concurrently": "^6.2.1",
     "cross-env": "^7.0.3",
     "eslint": "^8.56.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^27.6.3",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-monorepo-cop": "^1.0.2",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-promise": "^6.1.1",
     "husky": "^9.0.7",
     "jest": "^29.7.0",
-    "lerna": "^4.0.0",
-    "lint-staged": "^13.2.2",
+    "lint-staged": "^15.2.10",
     "prettier": "^3.2.4",
     "shx": "^0.3.3",
-    "typescript": "~5.3.3"
+    "typescript": "~5.6.2"
   },
   "workspaces": [
     "agent/*",
@@ -71,13 +56,7 @@
     "js-path",
     "real-user-agents",
     "plugins/*",
-    "double-agent/collect",
-    "double-agent/collect-controller",
-    "double-agent/config",
-    "double-agent/analyze",
-    "double-agent/docs",
-    "double-agent/tls-server",
-    "double-agent/runner",
+    "double-agent/*",
     "double-agent/collect/plugins/*",
     "double-agent/analyze/plugins/*",
     "double-agent-stacks",
