@@ -1,11 +1,13 @@
 import '@ulixee/commons/lib/SourceMapSupport';
-import * as stableChromeVersions from '@ulixee/real-user-agents/data/stableChromeVersions.json';
+import { defaultBrowserEngine } from '@ulixee/default-browser-emulator';
 import IUserAgentConfig from '@double-agent/runner/interfaces/IUserAgentConfig';
 import * as Fs from 'fs';
 import { getExternalDataPath } from '../paths';
 
+const currentBrowserId = defaultBrowserEngine.id;
+const prevBrowserId = `${defaultBrowserEngine.name}-${defaultBrowserEngine.majorVersion - 1}-0`;
 const agents: IUserAgentConfig = {
-  browserIds: stableChromeVersions.slice(0, 2).map(x => x.id),
+  browserIds: [currentBrowserId, prevBrowserId],
 };
 
 Fs.writeFileSync(getExternalDataPath('userAgentConfig.json'), JSON.stringify(agents, null, 2));
